refactor: moved ProductTagsDataSources to a separate file
